requirements.txt
appdirs==1.4.4
attrs==22.2.0
beautifulsoup4==4.11.2
beniget==0.4.1
Brotli==1.0.9
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
colorzero==2.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
distro==1.8.0
fonttools==4.38.0
fs==2.4.16
gast==0.5.2
gpiozero==2.0
html5lib==1.1
idna==3.3
kiwisolver==0.0.0
lgpio==0.2.2.0
lxml==4.9.2
lz4==4.0.2+dfsg
matplotlib==3.6.3
mpmath==0.0.0
numpy==1.24.2
olefile==0.46
packaging==23.0
pigpio==1.78
Pillow==9.4.0
ply==3.11
protobuf==4.21.12
pyparsing==3.0.9
pyserial==3.5
python-apt==2.6.0
python-dateutil==2.8.2
pythran==0.11.0
pytz==2022.7.1
requests==2.28.1
RPi.GPIO==0.7.1a4
scipy==1.10.1
six==1.16.0
smbus2==0.4.2
soupsieve==2.3.2
spidev==3.5
ssh-import-id==5.10
sympy==1.11.1
toml==0.10.2
ufoLib2==0.14.0
urllib3==1.26.12
webencodings==0.5.1
